Add notes from 1409 checkins
Overall, my advice was approximately the same:

Figure out what they needed to get their stuff on Heroku
with a database, an admin page, and the data dynamically
rendered into their content.

Take that, prioritize it, show them what they can work on concurrently,
which is basically all of it, if you separate the integration from
the feature itself (ie 2 people work on 2 different features that
go together like rendering the data into the form based on what
the controller gives it, and load the data from the database. Then
to actually get the real feature working, someone needs to take the
data that was loaded from the db and connect it to the data that's
being rendered into the form, but if they see that as a separate step,
then they can work on each of these two things concurrently, and then
perform the integration at the end when they're both in place,
thus allowing them to work better and integrate more frequently).

A little bit of agile advice like every tiny little thing is value,
so merge all the fucking time and don't get in these situation where
you have all these branches that you're scared to merge b/c of the
inevitable conflict, with teammates waiting for an existing part of
one person's feature which isn't merged because some other part
isn't in place, which doesn't matter for the person needing the
feature. Also, saw a few things like multiple people impelmenting
the same feature, because they hadn't merged and didn't realize
that the other one was doing it, too.

A little bit of "no, you don't have to waste your time doing that
tedious task, I don't actually care that your site is an exact replica
of the original, I care that you have the opportunity to get experience
with databases, controllers, authentication, collaboration, iterative
development of features, etc"

The exercise is a medium, y'all, not an end!